requirements.txt
numpy==2.3.5
sounddevice==0.5.3
openai-whisper==20250625
pyaudio==0.2.14
pyperclip==1.11.0
torchvision==0.24.1
torchaudio==2.9.1
scipy==1.16.3
